+++ src/nv_timeline.py
@@ -16,20 +16,14 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 """
-from datetime import datetime
-import os
 from pathlib import Path
-from tkinter import filedialog
 from tkinter import ttk
 import webbrowser
 
 from nvlib.controller.plugin.plugin_base import PluginBase
-from nvlib.model.file.doc_open import open_document
-from nvlib.novx_globals import Error
-from nvlib.novx_globals import norm_path
-from nvtimelinelib.nvtimeline_globals import _
-from nvtimelinelib.tl_file import TlFile
+from nvtimeline.nvtimeline_locale import _
 import tkinter as tk
+from nvtimeline.tl_service import TlService
 
 
 class Plugin(PluginBase):
@@ -38,19 +32,12 @@ class Plugin(PluginBase):
     API_VERSION = '5.0'
     DESCRIPTION = 'Synchronize with Timeline'
     URL = 'https://github.com/peter88213/nv_timeline'
-    _HELP_URL = f'{_("https://peter88213.github.io/nvhelp-en")}/nv_timeline/'
+    HELP_URL = f'{_("https://peter88213.github.io/nvhelp-en")}/nv_timeline/'
 
     FEATURE = 'Timeline'
-    INI_FILENAME = 'nv_timeline.ini'
-    INI_FILEPATH = '.novx/config'
-    SETTINGS = dict(
-        section_label='Section',
-        section_color='170,240,160',
-        new_event_spacing='1'
-    )
-    OPTIONS = dict(
-        lock_on_export=False,
-    )
+
+    def create_novx(self):
+        self.timelineService.create_novx()
 
     def disable_menu(self):
         """Disable menu entries when no project is open.
@@ -68,6 +55,15 @@ class Plugin(PluginBase):
         self._ui.toolsMenu.entryconfig(self.FEATURE, state='normal')
         self._timelineButton.config(state='normal')
 
+    def export_from_novx(self):
+        self.timelineService.export_from_novx()
+
+    def import_to_novx(self):
+        self.timelineService.import_to_novx()
+
+    def info(self):
+        self.timelineService.info()
+
     def install(self, model, view, controller):
         """Add a submenu to the main menu.
         
@@ -84,38 +80,46 @@ class Plugin(PluginBase):
         super().install(model, view, controller)
 
         # Create a submenu in the Tools menu.
-        self._pluginMenu = tk.Menu(self._ui.toolsMenu, tearoff=0)
-        self._ui.toolsMenu.add_cascade(label=self.FEATURE, menu=self._pluginMenu)
+        self.pluginMenu = tk.Menu(self._ui.toolsMenu, tearoff=0)
+        self._ui.toolsMenu.add_cascade(label=self.FEATURE, menu=self.pluginMenu)
         self._ui.toolsMenu.entryconfig(self.FEATURE, state='disabled')
-        self._pluginMenu.add_command(label=_('Information'), command=self._info)
-        self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Create or update the timeline'), command=self._export_from_novx)
-        self._pluginMenu.add_command(label=_('Update the project'), command=self._import_to_novx)
-        self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Open Timeline'), command=self._launch_application)
+        self.pluginMenu.add_command(label=_('Information'), command=self.info)
+        self.pluginMenu.add_separator()
+        self.pluginMenu.add_command(label=_('Create or update the timeline'), command=self.export_from_novx)
+        self.pluginMenu.add_command(label=_('Update the project'), command=self.import_to_novx)
+        self.pluginMenu.add_separator()
+        self.pluginMenu.add_command(label=_('Open Timeline'), command=self.launch_application)
 
         # Add an entry to the "File > New" menu.
-        self._ui.newMenu.add_command(label=_('Create from Timeline...'), command=self._create_novx)
+        self._ui.newMenu.add_command(label=_('Create from Timeline...'), command=self.create_novx)
 
         # Add an entry to the Help menu.
-        self._ui.helpMenu.add_command(label=_('Timeline plugin Online help'), command=lambda: webbrowser.open(self._HELP_URL))
+        self._ui.helpMenu.add_command(label=_('Timeline plugin Online help'), command=self.open_help_page)
 
         #--- Configure the toolbar.
         self._configure_toolbar()
+
+        self.timelineService = TlService(model, view, controller, self.FEATURE)
+
+    def launch_application(self):
+        self.timelineService.launch_application()
 
     def lock(self):
         """Inhibit changes on the model.
         
         Overrides the superclass method.
         """
-        self._pluginMenu.entryconfig(_('Update the project'), state='disabled')
+        self.pluginMenu.entryconfig(_('Update the project'), state='disabled')
+
+    def open_help_page(self):
+        webbrowser.open(self.HELP_URL)
 
     def unlock(self):
         """Enable changes on the model.
         
         Overrides the superclass method.
         """
-        self._pluginMenu.entryconfig(_('Update the project'), state='normal')
+        self.pluginMenu.entryconfig(_('Update the project'), state='normal')
 
     def _configure_toolbar(self):
 
@@ -143,7 +147,7 @@ class Plugin(PluginBase):
             self._ui.toolbar.buttonBar,
             text=_('Open Timeline'),
             image=tlIcon,
-            command=self._launch_application
+            command=self.launch_application
             )
         self._timelineButton.pack(side='left')
         self._timelineButton.image = tlIcon
@@ -158,167 +162,4 @@ class Plugin(PluginBase):
             return
 
         Hovertip(self._timelineButton, self._timelineButton['text'])
-
-    def _create_novx(self):
-        """Create a novelibre project from a timeline."""
-        timelinePath = filedialog.askopenfilename(
-            filetypes=[(TlFile.DESCRIPTION, TlFile.EXTENSION)],
-            defaultextension=TlFile.EXTENSION,
-            )
-        if not timelinePath:
-            return
-
-        self._ctrl.close_project()
-        root, __ = os.path.splitext(timelinePath)
-        novxPath = f'{root}{self._mdl.nvService.get_novx_file_extension()}'
-        kwargs = self._get_configuration(timelinePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = TlFile(timelinePath, **kwargs)
-        target = self._mdl.nvService.new_novx_file(novxPath)
-
-        if os.path.isfile(target.filePath):
-            self._ui.set_status(f'!{_("File already exists")}: "{norm_path(target.filePath)}".')
-            return
-
-        message = ''
-        try:
-            source.novel = self._mdl.nvService.new_novel()
-            source.read()
-            target.novel = source.novel
-            target.write()
-        except Error as ex:
-            message = f'!{str(ex)}'
-        else:
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-            self._ctrl.open_project(filePath=target.filePath, doNotSave=True)
-        finally:
-            self._ui.set_status(message)
-
-    def _export_from_novx(self):
-        """Update or create a timeline from the novelibre project."""
-        if not self._mdl.prjFile:
-            return
-
-        self._ui.propertiesView.apply_changes()
-        self._ui.restore_status()
-        if not self._mdl.prjFile.filePath:
-            if not self._ctrl.save_project():
-                return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{TlFile.EXTENSION}'
-        if os.path.isfile(timelinePath):
-            action = _('update')
-        else:
-            action = _('create')
-        if self._mdl.isModified:
-            if not self._ui.ask_yes_no(_('Save the project and {} the timeline?').format(action)):
-                return
-
-            self._ctrl.save_project()
-        elif action == _('update') and not self._ui.ask_yes_no(_('Update the timeline?')):
-            return
-
-        kwargs = self._get_configuration(self._mdl.prjFile.filePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = self._mdl.nvService.new_novx_file(self._mdl.prjFile.filePath)
-        source.novel = self._mdl.nvService.new_novel()
-        target = TlFile(timelinePath, **kwargs)
-        target.novel = self._mdl.nvService.new_novel()
-        try:
-            source.read()
-            if os.path.isfile(target.filePath):
-                target.read()
-            target.write(source.novel)
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-        except Error as ex:
-            message = f'!{str(ex)}'
-        self._ui.set_status(message)
-
-    def _get_configuration(self, sourcePath):
-        """Return a dictionary with persistent configuration data."""
-        sourceDir = os.path.dirname(sourcePath)
-        if not sourceDir:
-            sourceDir = '.'
-        try:
-            homeDir = str(Path.home()).replace('\\', '/')
-            pluginCnfDir = f'{homeDir}/{self.INI_FILEPATH}'
-        except:
-            pluginCnfDir = '.'
-        iniFiles = [f'{pluginCnfDir}/{self.INI_FILENAME}', f'{sourceDir}/{self.INI_FILENAME}']
-        configuration = self._mdl.nvService.new_configuration(
-            settings=self.SETTINGS,
-            options=self.OPTIONS
-            )
-        for iniFile in iniFiles:
-            configuration.read(iniFile)
-        configData = {}
-        configData.update(configuration.settings)
-        configData.update(configuration.options)
-        return configData
-
-    def _import_to_novx(self):
-        """Update the novelibre project from a timeline."""
-        if not self._mdl.prjFile:
-            return
-
-        self._ui.restore_status()
-        self._ui.propertiesView.apply_changes()
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{TlFile.EXTENSION}'
-        if not os.path.isfile(timelinePath):
-            self._ui.set_status(_('!No {} file available for this project.').format(self.FEATURE))
-            return
-
-        if self._mdl.isModified and not self._ui.ask_yes_no(_('Save the project and update it?')):
-            return
-
-        self._ctrl.save_project()
-        kwargs = self._get_configuration(timelinePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = TlFile(timelinePath, **kwargs)
-        target = self._mdl.nvService.new_novx_file(self._mdl.prjFile.filePath, **kwargs)
-        message = ''
-        try:
-            target.novel = self._mdl.nvService.new_novel()
-            target.read()
-            source.novel = target.novel
-            source.read()
-            target.novel = source.novel
-            target.write()
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-            self._ctrl.open_project(filePath=self._mdl.prjFile.filePath, doNotSave=True)
-        except Error as ex:
-            message = f'!{str(ex)}'
-        self._ui.set_status(f'{message}')
-
-    def _info(self):
-        """Show information about the Timeline file."""
-        if not self._mdl.prjFile:
-            return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{TlFile.EXTENSION}'
-        if os.path.isfile(timelinePath):
-            try:
-                timestamp = os.path.getmtime(timelinePath)
-                if timestamp > self._mdl.prjFile.timestamp:
-                    cmp = _('newer')
-                else:
-                    cmp = _('older')
-                fileDate = datetime.fromtimestamp(timestamp).strftime('%c')
-                message = _('{0} file is {1} than the novelibre project.\n (last saved on {2})').format(self.FEATURE, cmp, fileDate)
-            except:
-                message = _('Cannot determine file date.')
-        else:
-            message = _('No {} file available for this project.').format(self.FEATURE)
-        self._ui.show_info(message, title=self.FEATURE)
-
-    def _launch_application(self):
-        """Launch Timeline with the current project."""
-        if self._mdl.prjFile:
-            timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{TlFile.EXTENSION}'
-            if os.path.isfile(timelinePath):
-                if self.OPTIONS['lock_on_export']:
-                    self._ctrl.lock()
-                open_document(timelinePath)
-            else:
-                self._ui.set_status(_('!No {} file available for this project.').format(self.FEATURE))