+++ docs/changelog.md
@@ -5,13 +5,14 @@
 ## Changelog
 
 
-### Version 5.0.1
+### Version 5.0.2
 
 Library update:
 - Refactor the code for better maintainability.
+- Fix a bug where project cannot be optionally locked on opening the timeline.
 
 API: 5.0
-Based on novelibre 5.0.12
+Based on novelibre 5.0.24
 
 ### Version 4.4.3